requirements.txt
img2pdf==0.3.6
Pillow==7.2.0
qrcode==6.1